+++ client-core/src/main/java/com/influxdb/exceptions/InfluxException.java
@@ -89,7 +89,7 @@ public class InfluxException extends RuntimeException {
     }
 
     /**
-     * Gets the reference code unique to the error type. If the reference code is not present than return "0".
+     * Gets the reference code unique to the error type. If the reference code is not present then return "0".
      *
      * @return reference code unique to the error type
      */
@@ -109,7 +109,7 @@ public class InfluxException extends RuntimeException {
 
     /**
      * Gets the HTTP status code of the unsuccessful response.
-     * If the response is not present than return "0".
+     * If the response is not present then return "0".
      *
      * @return HTTP status code
      */
@@ -124,7 +124,7 @@ public class InfluxException extends RuntimeException {
 
     /**
      * Gets the HTTP headers from the unsuccessful response.
-     * If the response is not present than return empty {@code Map}.
+     * If the response is not present then return empty {@code Map}.
      *
      * @return HTTP headers
      */